move offers router to file
