+++ src/services/lease-service/index.ts
@@ -1,6 +1,5 @@
 import KoaRouter from '@koa/router'
 import { Applicant, ApplicantStatus, Listing } from 'onecore-types'
-import { z } from 'zod'
 
 import {
   getContactByContactCode,
@@ -38,9 +37,6 @@ import {
   getDetailedApplicantInformation,
   sortApplicantsBasedOnRentalRules,
 } from './priority-list-service'
-
-import * as offerAdapter from './adapters/offer-adapter'
-import { parseRequestBody } from '../../middlewares/parse-request-body'
 
 interface CreateLeaseRequest {
   parkingSpaceId: string
@@ -536,31 +532,4 @@ export const routes = (router: KoaRouter) => {
       }
     }
   })
-
-  const createOfferRequestParams = z.object({
-    expiresAt: z.coerce.date(),
-    status: z.nativeEnum(offerAdapter.OfferStatus),
-    selectionSnapshot: z.any().array(),
-    listingId: z.coerce.number(),
-    applicantId: z.coerce.number(),
-  })
-
-  // TODO: Use response type
-  const createSuccessResponse = (data: unknown) => ({ data })
-
-  router.post(
-    '(.*)/offer',
-    parseRequestBody(createOfferRequestParams),
-    async (ctx) => {
-      try {
-        const offer = await offerAdapter.create(ctx.request.body)
-
-        ctx.status = 201
-        ctx.body = createSuccessResponse(offer)
-      } catch (err) {
-        console.log('Error creating offer: ', JSON.stringify(err, null, 2))
-        ctx.status = 500
-      }
-    }
-  )
 }
